Moved pin-manager
