+++ webpack.config.js
@@ -5,7 +5,6 @@ module.exports = {
     entry: [
         './js/src/main.js',
         './js/src/site-controls.js',
-        './js/src/pin-manager.js',
     ],
     plugins: [new CleanWebpackPlugin(['js/dist'])],
     devtool: 'inline-source-map',
